moving tests to right folder type
